requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.6
dj-static==0.0.6
Django==4.1.3
django-dotenv==1.4.2
filelock==3.8.0
fonttools==4.42.1
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.4
importlib-resources==6.0.1
install-jdk==0.3.0
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.1
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.1
packaging==23.1
pandas==1.5.2
Pillow==9.3.0
platformdirs==2.5.4
psycopg2==2.9.5
pycodestyle==2.9.1
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
static3==0.7.0
sympy==1.12
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing-extensions==4.7.1
tzdata==2022.6
urllib3==1.26.12
virtualenv==20.16.7
virtualenvwrapper-win==1.2.7
whitenoise==6.2.0
zipp==3.16.2
